requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
joblib==1.5.2
pandas==2.3.3
numpy==2.3.4
matplotlib==3.10.7
scikit-learn==1.7.2
requests==2.32.5
mysql-connector-python==9.5.0
plotly==5.24.1
streamlit==1.40.2
#env python 64 bits <---importante